♻️ Create OpaqueModelRenderer class
This class will be used for rendering opaque models, such as
rectangular prisms (and eventually shapes made out of such prisms).
With the change, boxes can be rendered by providing two corners.
It also moves a lot of the rendering code into the class, such as
shaders and pipeline creation.