🔊 Introduce tracing facility
This change introduces a basic task-local tracing facility. At the
moment it just prints the message and doesn't do any filtering based
on level.

I chose against using the swift-log library (or something similar)
because it didn't really serve my purpose, which was fine-grained
tracing of internal, developer-targeted messages. Since I just want
simple output to the terminal at the moment, this will
suffice. There's also nothing stopping a logging library from being
added as the "backend" for this facility in the future.